Tidy up view handling
