+++ assets/js/blocks/cart-checkout-shared/view-switcher/utils.ts
@@ -1,4 +1,9 @@
 /**
+ * External dependencies
+ */
+import { select } from '@wordpress/data';
+
+/**
  * Internal dependencies
  */
 import type { View } from './types';
@@ -8,4 +13,39 @@ export const getView = (
 	views: View[]
 ): View | undefined => {
 	return views.find( ( view ) => view.view === viewName );
+};
+
+const defaultView = {
+	views: [],
+	currentView: '',
+	viewClientId: '',
+};
+
+export const findParentBlockEditorViews = (
+	clintId: string
+): {
+	views: View[];
+	currentView: string;
+	viewClientId: string;
+} => {
+	const { getBlockAttributes, getBlockRootClientId } =
+		select( 'core/block-editor' );
+	const rootId = getBlockRootClientId( clintId ) as string;
+	const rootAttributes = rootId ? getBlockAttributes( rootId ) : null;
+
+	if ( ! rootAttributes ) {
+		return defaultView;
+	}
+
+	if ( rootAttributes.editorViews !== undefined ) {
+		return {
+			views: rootAttributes.editorViews,
+			currentView:
+				rootAttributes.currentView ||
+				rootAttributes.editorViews[ 0 ].view,
+			viewClientId: rootId,
+		};
+	}
+
+	return findParentBlockEditorViews( rootId );
 };
