+++ package.json
@@ -122,7 +122,6 @@
 		"@types/react-dom": "18.0.9",
 		"@types/wordpress__block-editor": "6.0.6",
 		"@types/wordpress__blocks": "11.0.7",
-		"@types/wordpress__compose": "4.0.1",
 		"@types/wordpress__core-data": "^2.4.5",
 		"@types/wordpress__data": "^6.0.1",
 		"@types/wordpress__data-controls": "2.2.0",
